feat: add with-headers util
BREAKING CHANGE: `withAuth` is now exported from lib/withHeader
