+++ packages/ts-endpoint/src/Endpoint/__typetests__/withHeaders.typespec.ts
@@ -1,6 +1,6 @@
 import * as t from 'io-ts';
 import { Endpoint } from '..';
-import { withAuth } from '../withAuth';
+import { withAuth, withHeaders } from '../withHeaders';
 
 const endpoint = Endpoint({
   Input: {
@@ -37,3 +37,20 @@ authenticatedNoHeaderEndpoint.Input.Headers.props.authorization;
 
 // @dts-jest:fail:snap when endpoint is not wrapped you must not pass the header
 noHeaderEndpoint.Input.Headers?.authorization;
+
+
+const withHeadersEndpoint = withHeaders({
+  foo: t.string
+})(endpoint)
+
+// @dts-jest:pass:snap when endpoint is wrapped other headers are still present
+withHeadersEndpoint.Input.Headers.props.id;
+// @dts-jest:pass:snap when endpoint is wrapped foo header is present
+withHeadersEndpoint.Input.Headers.props.foo;
+
+const withHeadersNoHeaderEndpoint = withHeaders({
+  foo: t.string
+})(noHeaderEndpoint)
+
+// @dts-jest:pass:snap when endpoint with no previous headers is wrapped, foo header is present
+withHeadersNoHeaderEndpoint.Input.Headers.props.foo;
